+++ packages/spec/lib/src/map_expect.dart
@@ -1,29 +1,23 @@
 part of 'expect.dart';
 
-/// Matchers for [Map]s.
-extension MapExpectation<Return> on ExpectationBase<Object?, Return, Map> {
-  /// A matcher that matches if the argument is a map and has the [expected] key.
-  /// ```dart
-  /// expect({'hello': 'world'}).toHaveKey('hello');
-  /// ```
-  Return toHaveKey(Object? expected) {
-    return toContain(expected);
-  }
+/// Matchers for Maps
+extension MapExpectationX<Return, KeyParam, ValueParam> on ExpectationBase<
+    Map<KeyParam, ValueParam>, Return, Map<KeyParam, ValueParam>> {
+  /// Returns a matcher which matches maps containing the key-value pair
+  /// with [key] => [value].
+  Return containsPair(KeyParam key, ValueParam value) =>
+      runMatcher(dart_test.containsPair(key, value));
 
-  /// Returns a matcher which matches maps containing the given [value].
-  /// ```dart
-  /// expect({'hello': 'world'}).toHaveValue('world');
-  /// ```
-  Return toHaveValue(Object? value) {
-    return runMatcher(dart_test.containsValue(value));
-  }
+  /// Returns a matcher which matches maps containing the [key]
+  /// with any value.
+  Return containsKey(KeyParam key) => toContain(key);
 
   /// Returns a matcher which matches maps containing the key-value pair
   /// with [key] => [valueOrMatcher].
   /// ```dart
   /// expect({'hello': 'world'}).toHaveKeyValuePair('hello', 'world');
   /// ```
-  Return toHaveKeyValuePair(Object? key, Object? valueOrMatcher) {
-    return runMatcher(dart_test.containsPair(key, valueOrMatcher));
+  Return containsValue(ValueParam? valueOrMatcher) {
+    return runMatcher(dart_test.containsValue(valueOrMatcher));
   }
 }