+++ packages/spec/test/spec_test.dart
@@ -190,30 +190,61 @@ void main() {
     });
   });
 
-  group('Map matchers', () {
-    test('toHaveKey', () {
-      expect({'hello': 'world'}).toHaveKey('hello');
-
-      t.expect(() => expect({'hello': 'world'}).toHaveKey('world'),
-          throwsTestFailure);
-    });
-
-    test('toHaveValue', () {
-      expect({'hello': 'world'}).toHaveValue('world');
-
-      t.expect(() => expect({'hello': 'world'}).toHaveValue('hello'),
-          throwsTestFailure);
-    });
-
-    test('toHaveKeyValuePair', () {
-      expect({'hello': 'world'}).toHaveKeyValuePair('hello', 'world');
+  group('list', () {
+    test('containsAll', () {
+      expect([1, 2, 3]).containsAll([1, 2]);
 
       t.expect(
-          () => expect({'hello': 'world'}).toHaveKeyValuePair('bye', 'world'),
-          throwsTestFailure);
+        () => expect([1, 2, 3]).containsAll([1, 2, 4]),
+        throwsTestFailure,
+      );
+    });
+
+    test('not contains', () {
+      expect([1, 2, 3]).not.containsAll([1, 2, 4]);
+
       t.expect(
-          () => expect({'hello': 'world'}).toHaveKeyValuePair('hello', 'dart'),
-          throwsTestFailure);
+        () => expect([1, 2, 3]).not.containsAll([1, 2]),
+        throwsTestFailure,
+      );
+    });
+
+    test('contains', () {
+      expect([1, 2, 3]).contains(1);
+
+      t.expect(
+        () => expect([1, 2, 3]).contains(4),
+        throwsTestFailure,
+      );
+    });
+  });
+
+  group('map', () {
+    test('containsPair', () {
+      expect({'a': 1, 'b': 2}).containsPair('a', 1);
+
+      t.expect(
+        () => expect({'a': 1, 'b': 2}).containsPair('a', 2),
+        throwsTestFailure,
+      );
+    });
+
+    test('containsKey', () {
+      expect({'a': 1, 'b': 2}).containsKey('a');
+
+      t.expect(
+        () => expect({'a': 1, 'b': 2}).containsKey('c'),
+        throwsTestFailure,
+      );
+    });
+
+    test('containsValue', () {
+      expect({'a': 1, 'b': 2}).containsValue(1);
+
+      t.expect(
+        () => expect({'a': 1, 'b': 2}).containsValue(3),
+        throwsTestFailure,
+      );
     });
   });
 }